Switching over to NBT
Old data format, while it can easily be editted in a text editor, never actually worked well (that's more my fault really). Decided to switch over to NBT since it's a format I'm familiar with, it's simple to use, and while it can't be as easily edited as JSON or any other text format, there are plenty of editors out there that make reading/editing it really easy as long as we follow its specifications (documented here: http://wiki.vg/NBT#Specification).